refactor(chat): extract chat logic into event-chat component so it can be re-used in group chat
Refs #156
